requirements.txt
Django==4.0.5
openpyxl==3.0.10
djangorestframework==3.13.1
python-dotenv==0.20.0